requirements.txt
Pillow>=9.0
tqdm>=4.0